+++ source/main/utils/setupLogging.js
@@ -4,12 +4,14 @@ import moment from 'moment';
 import ensureDirectoryExists from './ensureDirectoryExists';
 import { pubLogsFolderPath, APP_NAME } from '../config';
 
+const isTest = process.env.NODE_ENV === 'test';
+
 export const setupLogging = () => {
   const logFilePath = path.join(pubLogsFolderPath, APP_NAME + '.log');
   ensureDirectoryExists(pubLogsFolderPath);
 
-  log.transports.console.level = false;
-  log.transports.rendererConsole.level = 'warn';
+  log.transports.console.level = isTest ? 'error' : false;
+  log.transports.rendererConsole.level = 'error';
   log.transports.file.level = 'debug';
   log.transports.file.maxSize = 20 * 1024 * 1024;
   log.transports.file.file = logFilePath;