+++ source/renderer/app/api/ada/types.js
@@ -4,6 +4,12 @@
 export type AdaAssurance = 'CWANormal' | 'CWAStrict';
 export type AdaTransactionCondition = 'CPtxApplying' | 'CPtxInBlocks' | 'CPtxWontApply' | 'CPtxNotTracked';
 export type AdaWalletRecoveryPhraseResponse = Array<string>;
+export type AdaWalletCertificateAdditionalMnemonicsResponse = Array<string>;
+export type AdaWalletCertificateRecoveryPhraseResponse = Array<string>;
+export type AdaWalletRecoveryPhraseFromCertificateResponse = Array<string>;
+export type GetWalletCertificateAdditionalMnemonicsResponse = Array<string>;
+export type GetWalletCertificateRecoveryPhraseResponse = Array<string>;
+export type GetWalletRecoveryPhraseFromCertificateResponse = Array<string>;
 
 export type AdaSyncProgressResponse = {
   _spLocalCD: {