+++ source/renderer/app/components/sidebar/Sidebar.js
@@ -8,6 +8,7 @@ import SidebarCategory from './SidebarCategory';
 import SidebarWalletsMenu from './wallets/SidebarWalletsMenu';
 import WalletAddDialog from '../../components/wallet/WalletAddDialog';
 import BugReportDialog from '../../components/profile/bug-report/BugReportDialog';
+import InstructionsDialog from '../wallet/paper-wallet-certificate/InstructionsDialog';
 import supportIcon from '../../assets/images/sidebar/bug-report-ic.inline.svg';
 import type { SidebarWalletType } from '../../stores/SidebarStore';
 
@@ -43,8 +44,7 @@ export default class Sidebar extends Component<Props> {
   render() {
     const {
       menus, categories, activeSidebarCategory,
-      isShowingSubMenus, onCategoryClicked,
-      openDialogAction, isDialogOpen,
+      isShowingSubMenus, openDialogAction, isDialogOpen,
     } = this.props;
     let subMenu = null;
 
@@ -79,7 +79,7 @@ export default class Sidebar extends Component<Props> {
                 className={categoryClassName}
                 icon={category.icon}
                 active={activeSidebarCategory === category.route}
-                onClick={() => onCategoryClicked(category.route)}
+                onClick={() => this.handleClick(category.name, category.route)}
               />
             );
           })}
@@ -88,18 +88,25 @@ export default class Sidebar extends Component<Props> {
             className="supportRequest"
             icon={supportIcon}
             active={isDialogOpen(BugReportDialog)}
-            onClick={this.handleSupportRequestClick}
+            onClick={() => this.handleClick('BUG_REPORT')}
           />
-
         </div>
         {subMenu}
       </div>
     );
   }
 
-  handleSupportRequestClick = () => {
-    this.props.openDialogAction({
-      dialog: BugReportDialog
-    });
-  }
+  handleClick = (categoryName: string, categoryRoute?: string) => {
+    if (categoryName === 'PAPER_WALLET_CREATE_CERTIFICATE') {
+      this.props.openDialogAction({
+        dialog: InstructionsDialog
+      });
+    } else if (categoryName === 'BUG_REPORT') {
+      this.props.openDialogAction({
+        dialog: BugReportDialog
+      });
+    } else {
+      this.props.onCategoryClicked(categoryRoute);
+    }
+  };
 }