+++ source/renderer/app/containers/wallet/dialogs/WalletRestoreDialogContainer.js
@@ -13,29 +13,50 @@ export default class WalletRestoreDialogContainer extends Component<Props> {
 
   static defaultProps = { actions: null, stores: null, children: null, onClose: () => {} };
 
-  onSubmit = (values: { recoveryPhrase: string, walletName: string, walletPassword: ?string }) => {
+  onSubmit = (values: {
+    recoveryPhrase: string,
+    walletName: string,
+    walletPassword: ?string,
+    type?: string,
+  }) => {
     this.props.actions[environment.API].wallets.restoreWallet.trigger(values);
   };
 
   onCancel = () => {
     this.props.onClose();
+    this.resetRequests();
+  };
+
+  resetRequests = () => {
     // Restore request should be reset only in case restore is finished/errored
-    const { restoreRequest } = this._getWalletsStore();
-    if (!restoreRequest.isExecuting) restoreRequest.reset();
+    const wallets = this._getWalletsStore();
+    const { restoreRequest } = wallets;
+    if (!restoreRequest.isExecuting) {
+      restoreRequest.reset();
+      if (environment.isAdaApi()) {
+        wallets.getWalletRecoveryPhraseFromCertificateRequest.reset();
+      }
+    }
   };
 
   render() {
     const wallets = this._getWalletsStore();
-    const { restoreRequest } = wallets;
+    const { restoreRequest, isValidMnemonic } = wallets;
+
+    const error = environment.isAdaApi()
+      ? (restoreRequest.error || wallets.getWalletRecoveryPhraseFromCertificateRequest.error)
+      : restoreRequest.error;
 
     return (
       <WalletRestoreDialog
-        mnemonicValidator={mnemonic => wallets.isValidMnemonic(mnemonic)}
+        mnemonicValidator={mnemonic => isValidMnemonic(mnemonic)}
+        showCertificateRestore={environment.isAdaApi()}
         suggestedMnemonics={validWords}
         isSubmitting={restoreRequest.isExecuting}
         onSubmit={this.onSubmit}
         onCancel={this.onCancel}
-        error={restoreRequest.error}
+        onChoiceChange={environment.isAdaApi() ? this.resetRequests : null}
+        error={error}
       />
     );
   }